fix setIsNightMode
