+++ src/App.tsx
@@ -21,6 +21,15 @@ function App() {
     }
   }, [background, isNightMode]);
 
+  useEffect(() => {
+    const now = new Date();
+    const hour = now.getHours();
+    if (hour >= 18 || hour < 6) {
+      setIsNightMode(true);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
     <div data-component="App" key={appKey} style={{ opacity: opacity }}>
       <GptImagesProvider>
