+++ src/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
@@ -9,15 +9,9 @@ const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
 );
 
-function ErrorComponent() {
-  throw new Error('This is a test error');
-  return <div>ErrorComponent</div>;
-}
-
 root.render(
   <React.StrictMode>
     <ErrorBoundary>
-      {/* <ErrorComponent /> */}
       <App />
     </ErrorBoundary>
   </React.StrictMode>,
