+++ src/pages/ConsolePage.scss
@@ -398,9 +398,9 @@
       &.deepseek {
         background-image: url('../static/deepseek.svg');
         background-repeat: no-repeat; 
-        background-size: cover;
+        background-size: 30px;
         background-position: center center;
-        background-color: rgba(255, 255, 255, 0.5); 
+        background-color: rgba(255, 255, 255); 
         background-clip: content-box; 
       }
